Fix library `.so` detection
The old way of creating the list of shared libraries of the llvm
installation was error-prone (it used `glob`) and did not work in all
installations (e.g. `libLLVM-18.so`, split installations with
`libLLVMCore`, `libLLVMSupport`). First use `llvm-config --shared-mode`
and if that is `shared` use `llvm-config --libnames` to gather all the
needed libraries. Keep the old system as there are situations where the
`--shared-mode` is reported as `static` but the package still has
`.so`s.